Use common interface for ERC721 details
