+++ packages/shared/src/lib/core/nfts/interfaces/index.ts
@@ -1,9 +1,9 @@
 export * from './download-queue-nft-item.interface'
 export * from './erc721-contract-metadata.interface'
 export * from './erc721-token-metadata.interface'
-export * from './irc27-attribute.interface'
 export * from './irc27-metadata.interface'
 export * from './nft.interface'
+export * from './nft-attribute.interface'
 export * from './nft-download-status.interface'
 export * from './persisted-nft.interface'
 export * from './persisted-nfts.interface'
