+++ packages/shared/src/lib/core/nfts/interfaces/nft-attribute.interface.ts
@@ -1,6 +1,6 @@
 // This interface does not follow the coding conventions as it is
-// following TIP27 metadata standards
-export interface IIrc27Attribute {
+// following TIP27 and ERC721 metadata standards
+export interface INftAttribute {
     trait_type: string
     value: string | number
 }
